+++ flight/Modules/PathPlanner/pathplanner.c
@@ -371,7 +371,7 @@ static void createPathBox()
 	pathPlannerStatus.NumberOfWaypoints = 7;
 	PathPlannerStatusSet(&pathPlannerStatus);
 
-	for (int i=UAVObjGetNumInstances(WaypointHandle()); i<pathPlannerStatus.NumberOfWaypoints; i++) {
+	for (uint16_t i=UAVObjGetNumInstances(WaypointHandle()); i<pathPlannerStatus.NumberOfWaypoints; i++) {
 		WaypointCreateInstance();
 	}
 
@@ -429,7 +429,7 @@ static void createPathStar()
 	pathPlannerStatus.NumberOfWaypoints = 8;
 	PathPlannerStatusSet(&pathPlannerStatus);
 	
-	for (int i=UAVObjGetNumInstances(WaypointHandle()); i<pathPlannerStatus.NumberOfWaypoints; i++) {
+	for (uint16_t i=UAVObjGetNumInstances(WaypointHandle()); i<pathPlannerStatus.NumberOfWaypoints; i++) {
 		WaypointCreateInstance();
 	}
 	
@@ -445,7 +445,7 @@ static void createPathStar()
 	WaypointInstSet(0, &waypoint);
 	
 	// Make five sides of star, plus one extra path to get to the start of the star
-	for (int i=1; i<7; i++) {
+	for (uint16_t i=1; i<7; i++) {
 		waypoint.Position[0] = 35*scale*cosf(theta);
 		waypoint.Position[1] = 35*scale*sinf(theta);
 		waypoint.Position[2] = -50;
@@ -472,7 +472,7 @@ static void createPathLogo()
 	pathPlannerStatus.NumberOfWaypoints = 42;
 	PathPlannerStatusSet(&pathPlannerStatus);
 
-	for (int i=UAVObjGetNumInstances(WaypointHandle()); i<pathPlannerStatus.NumberOfWaypoints; i++) {
+	for (uint16_t i=UAVObjGetNumInstances(WaypointHandle()); i<pathPlannerStatus.NumberOfWaypoints; i++) {
 		WaypointCreateInstance();
 	}
